feat(ui): move railroad logos and roadnameLogos to @repo/ui
Copy all railroad logo assets from the throttle app into @repo/ui, create
the Roster barrel with ROADNAME_MEDIA/getRoadnameMedia/RoadnameMedia, add
roster exports to the ui package index, and replace the throttle source
with a thin re-export so existing consumers are unaffected.
