+++ packages/ui/src/Roster/index.ts
@@ -1,0 +1,2 @@
+export { ROADNAME_MEDIA, getRoadnameMedia } from './roadnameLogos'
+export type { RoadnameMedia } from './roadnameLogos'
