+++ packages/ui/src/index.ts
@@ -5,6 +5,10 @@ export { default as Logo } from './Logo.vue'
 export { default as SplashPage } from './SplashPage.vue'
 export { default as NotFoundPage } from './NotFoundPage.vue'
 export { default as LocoAvatar } from './LocoAvatar.vue'
+
+// Roster components
+export { ROADNAME_MEDIA, getRoadnameMedia } from './Roster'
+export type { RoadnameMedia } from './Roster'
 export { default as ConsistIndicator } from './Consist/ConsistIndicator.vue'
 export { default as ConsistEditor } from './Consist/ConsistEditor.vue'
 export { default as Functions } from './Functions/Functions.vue'
